+++ code-review.el
@@ -74,6 +74,11 @@
   :group 'code-review
   :type 'integer)
 
+(defcustom code-review-download-dir "/tmp/code-review/"
+  "Directory where code review will download binary files."
+  :type 'string
+  :group 'code-review)
+
 (defcustom code-review-log-file (expand-file-name
                                  "code-review-error.log"
                                  user-emacs-directory)
@@ -221,21 +226,21 @@ If you want to display a minibuffer MSG in the end."
 
 (cl-defmethod code-review--internal-build ((_github code-review-github-repo) progress res &optional buff-name msg)
   "Helper function to build process for GITHUB based on the fetched RES informing PROGRESS."
-  ;; 2. save raw diff data
-  (progress-reporter-update progress 3)
-  (code-review-db--pullreq-raw-diff-update
-   (code-review-utils--clean-diff-prefixes
-    (a-get (-first-item res) 'message)))
+  (let* ((raw-infos (a-get-in (-second-item res) (list 'data 'repository 'pullRequest))))
+    ;; 2. save raw diff data
+    (progress-reporter-update progress 3)
+    (code-review-db--pullreq-raw-diff-update
+     (code-review-utils--clean-diff-prefixes
+      (a-get (-first-item res) 'message)))
 
-  ;; 2.1 save raw info data e.g. data from GraphQL API
-  (progress-reporter-update progress 4)
-  (code-review-db--pullreq-raw-infos-update
-   (a-get-in (-second-item res) (list 'data 'repository 'pullRequest)))
+    ;; 2.1 save raw info data e.g. data from GraphQL API
+    (progress-reporter-update progress 4)
+    (code-review-db--pullreq-raw-infos-update raw-infos)
 
-  ;; 2.2 trigger renders
-  (progress-reporter-update progress 5)
-  (code-review--trigger-hooks buff-name msg)
-  (progress-reporter-done progress))
+    ;; 2.2 trigger renders
+    (progress-reporter-update progress 5)
+    (code-review--trigger-hooks buff-name msg)
+    (progress-reporter-done progress)))
 
 (cl-defmethod code-review--internal-build ((_gitlab code-review-gitlab-repo) progress res &optional buff-name msg)
   "Helper function to build process for GITLAB based on the fetched RES informing PROGRESS."