+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [#49](https://github.com/wandersoncferreira/code-review/pull/46): promote comment to new issue. Github only.
 - [#85](https://github.com/wandersoncferreira/code-review/pull/85): remove comment and feedback section with `k`
 - [#86](https://github.com/wandersoncferreira/code-review/pull/86): allow users to define function to open Code Review buffer.
+- [#90](https://github.com/wandersoncferreira/code-review/pull/90): support visit binary files on Dired (`RET`) or Remote (`C-c C-v`)
 
 # v0.0.2